+++ app/src/main/java/com/adammcneilly/wom/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.adammcneilly.wom.theme
 
 import androidx.compose.ui.graphics.Color
