+++ settings.gradle.kts
@@ -5,5 +5,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "template"
+rootProject.name = "Wise Old Man"
 include(":app")
